make validateAccountName a global function
