+++ src/components/ProfileInfo/helper.ts
@@ -1,20 +1,4 @@
-import { Validator } from "@/hooks/useInput";
 import { UploadApiOptions, UploadApiResponse } from "cloudinary";
-
-// Account name is valid if it's unique in database
-export const validateAccountName: Validator = {
-	name: "validate account name",
-	async validatorFn(value: string) {
-		const response = await fetch(
-			`/api/validate-account-name?accountName=${value}`
-		);
-		const data = await response.json();
-		return {
-			valid: data.valid,
-			message: `${value} is taken`,
-		};
-	},
-};
 
 export function parseAccountName(accountName: string) {
 	return accountName?.toString().replaceAll(" ", "");
