+++ src/helpers/global_helpers.ts
@@ -1,3 +1,20 @@
+import { Validator } from "@/hooks/useInput";
+
+// Account name is valid if it's unique in database
+export const validateAccountName: Validator = {
+	name: "validate account name",
+	async validatorFn(value: string) {
+		const response = await fetch(
+			`/api/validate-account-name?accountName=${value}`
+		);
+		const data = await response.json();
+		return {
+			valid: data.valid,
+			message: `${value} is taken`,
+		};
+	},
+};
+
 export function classNames(...classNames: Array<boolean | string>) {
 	return classNames.filter(Boolean).join(" ");
 }
