+++ src/pages/create-profile/index.tsx
@@ -1,7 +1,5 @@
-import {
-	parseAccountName,
-	validateAccountName,
-} from "@/components/ProfileInfo/helper";
+import { parseAccountName } from "@/components/ProfileInfo/helper";
+import { validateAccountName } from "@/helpers/global_helpers";
 import useInput from "@/hooks/useInput";
 import { Button, Loader, Skeleton, TextInput, Title } from "@mantine/core";
 import { useSession } from "next-auth/react";
